terminate/requirements.txt
python-docx==1.1.0
python-colored-print
pypiwin32
keyboard
pyperclip
requests
ipython
pypdf==4.2.0
PyPDF2
pyyaml